requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
plotly>=5.13.0
st-gsheets-connection<=0.1.0
streamlit-js-eval>=0.1.0
Pillow>=10.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pytz>=2023.3
python-dotenv>=1.0.0
matplotlib>=3.8.0
fpdf2>=2.7.0
supabase